pyproject.toml
[tool.poetry]
name = "aicamp"
version = "0.1.0"
description = "Ai camp hackathon starter"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.35.0"
langchain = "^0.2.1"
certifi = "^2024.2.2"
pymongo = "^4.7.2"
openai = "^1.30.5"
python-dotenv = "^1.0.1"
watchdog = "^4.0.1"
langchain-community = "^0.2.1"
langchain-openai = "^0.1.8"
requests = "^2.32.3"
bs4 = "^0.0.2"
langchain-upstage = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
